+++ Plantory/Plantory/Assets.xcassets/Home/clover.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "mdi_clover.pdf",
+      "filename" : "clover.pdf",
       "idiom" : "universal"
     }
   ],
