+++ Plantory/Plantory/Assets.xcassets/Home/currentday.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "플러스 1.pdf",
+      "filename" : "currentday.pdf",
       "idiom" : "universal"
     }
   ],
