+++ Plantory/Plantory/Common/Enum/Home/EmotionColors.swift
